Refactor extract welcome message to user interface module
User input capture and processing, as well as the output messages
intended for the user form part of the "user interface".

Different user interfaces could (eventually) be provided to wrap
the core library. Even though that is not the intention, extracting
a module for the current (command-line)  interface will separate
concerns and allow for easier unit testing.